services/graphgen/pyproject.toml
[project]
name = "graphgen"
version = "0.1.0"
description = "Graph generation and ingestion service"
requires-python = ">=3.11"
dependencies = [
    # API
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    # Graph & databases
    "neo4j>=5.21.0",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.3.0",
    "networkx>=3.3",
    # ML / NLP
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "torch>=2.2.0",
    "sentence-transformers>=3.0.0",
    "spacy>=3.7.0",
    "nltk>=3.8.0",
    "leidenalg>=0.10.0",
    "gliner>=0.2.0",
    # LangChain
    "langchain-core>=0.2.0",
    "langchain-text-splitters>=0.2.0",
    "langchain-experimental>=0.0.60",
    "langchain-groq>=0.1.0",
    # Utilities
    "jellyfish>=1.0.0",
    # Document parsing
    "pymupdf4llm>=0.0.17",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    "openpyxl>=3.1.0",
    "trafilatura>=1.12.0",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
